Make rich toggleable and seperate out a new environment utility file (#779)
* Toggleable rich

* Refactor into environment utils
