+++ src/accelerate/utils/__init__.py
@@ -19,6 +19,7 @@ from .dataclasses import (
     SageMakerDistributedType,
     TensorInformation,
 )
+from .environment import get_int_from_env, parse_choice_from_env, parse_flag_from_env
 from .imports import (
     get_ccl_version,
     is_aim_available,
